IEEE Standard 754: Floating point numbers
This effort started in 1976 under Intel’s sponsorship with
the design of the 8087, a chip that provided floating-point
support for the 8086 processor. Intel hired William Kahan,
a professor at the University of California, Berkeley, as
a consultant to help design a floating-point standard for
its future processors. They allowed Kahan to join forces with
a committee generating an industry-wide standard under the
auspices of the Institute of Electrical and Elec- tronics
Engineers (IEEE). The committee ultimately adopted a standard
close to the one Kahan had devised for Intel. Nowadays,
virtually all computers support what has become known as
IEEE floating point. This has greatly improved the portability
of scientific application programs across different machines.